优化修改文章接口，同时给删除文章接口加上文章是否存在判断

+++ routers/api/v1/article.go
@@ -135,36 +135,55 @@ func AddArticle(c *gin.Context) {
 
 // 修改文章
 func EditArticle(c *gin.Context) {
+	valid := validation.Validation{}
+
+	id := com.StrTo(c.Param("id")).MustInt()
 	tagId := com.StrTo(c.Query("tag_id")).MustInt()
 	title := c.Query("title")
 	desc := c.Query("desc")
 	content := c.Query("content")
-	createdBy := c.Query("created_by")
-	state := com.StrTo(c.DefaultQuery("state", "0")).MustInt()
+	modifiedBy := c.Query("modified_by")
 
-	valid := validation.Validation{}
-	valid.Min(tagId, 1, "tag_id").Message("标签ID必须大于0")
-	valid.Required(title, "title").Message("标题不能为空")
-	valid.Required(desc, "desc").Message("简述不能为空")
-	valid.Required(content, "content").Message("内容不能为空")
-	valid.Required(createdBy, "created_by").Message("创建人不能为空")
-	valid.Range(state, 0, 1, "state").Message("状态只允许0或1")
+	var state int = -1
+	if arg := c.Query("state"); arg != "" {
+		state = com.StrTo(arg).MustInt()
+		valid.Range(state, 0, 1, "state").Message("状态只允许0或1")
+	}
+
+	valid.Min(id, 1, "id").Message("ID必须大于0")
+	valid.MaxSize(title, 100, "title").Message("标题最长为100字符")
+	valid.MaxSize(desc, 255, "desc").Message("简述最长为255字符")
+	valid.MaxSize(content, 65535, "content").Message("内容最长为65535字符")
+	valid.Required(modifiedBy, "modified_by").Message("修改人不能为空")
+	valid.MaxSize(modifiedBy, 100, "modified_by").Message("修改人最长为100字符")
 
 	code := e.INVALID_PARAMS
 	if ! valid.HasErrors() {
-		if models.ExistTagById(tagId) {
-			data := make(map[string]interface {})
-			data["tag_id"] = tagId
-			data["title"] = title
-			data["desc"] = desc
-			data["content"] = content
-			data["created_by"] = createdBy
-			data["state"] = state
+		if models.ExistArticleById(id) {
+			if models.ExistTagById(tagId) {
+				data := make(map[string]interface {})
+				if tagId > 0 {
+					data["tag_id"] = tagId
+				}
+				if title != "" {
+					data["title"] = title
+				}
+				if desc != "" {
+					data["desc"] = desc
+				}
+				if content != "" {
+					data["content"] = content
+				}
 
-			models.AddArticle(data)
-			code = e.SUCCESS
+				data["modified_by"] = modifiedBy
+
+				models.EditArticle(id, data)
+				code = e.SUCCESS
+			} else {
+				code = e.ERROR_NOT_EXIST_TAG
+			}
 		} else {
-			code = e.ERROR_NOT_EXIST_TAG
+			code = e.ERROR_NOT_EXIST_ARTICLE
 		}
 	} else {
 		for _, err := range valid.Errors {
@@ -175,7 +194,7 @@ func EditArticle(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"code" : code,
 		"msg" : e.GetMsg(code),
-		"data" : make(map[string]interface{}),
+		"data" : make(map[string]string),
 	})
 }
 
@@ -188,8 +207,13 @@ func DeleteArticle(c *gin.Context) {
 
 	code := e.INVALID_PARAMS
 	if !valid.HasErrors() {
-		code = e.SUCCESS
-		models.DeleteArticle(id)
+		if models.ExistArticleById(id) {
+			code = e.SUCCESS
+			models.DeleteArticle(id)
+		} else {
+			code = e.ERROR_NOT_EXIST_ARTICLE
+		}
+
 	} else {
 		for _, err := range valid.Errors {
 			log.Printf("err.key: %s, err.message: %s", err.Key, err.Message)
